+++ ecos/lca/l4_optimization/__init__.py
@@ -4,6 +4,7 @@ from .attribution import CTA_L4_Backend, CausalEffect, LCAAttribution
 from .ca_state_machine import CAStateMachine
 from .linucb import BanditConfig, LinUCB
 from .pomdp import POMDPConfig, POMDPPolicy
+from .pomdp_learner import RewardPosterior, TransitionPosterior
 from .policy_learner import LCAPolicyLearner
 from .thompson import ThompsonConfig, ThompsonSampling
 
@@ -15,6 +16,8 @@ __all__ = [
     "ThompsonConfig",
     "POMDPPolicy",
     "POMDPConfig",
+    "TransitionPosterior",
+    "RewardPosterior",
     "LCAPolicyLearner",
     "LCAAttribution",
     "CausalEffect",